Refactor: use MainDriver.COLLECTION_NAME
