+++ bosk-mongo/src/main/java/io/vena/bosk/drivers/mongo/SingleDocumentMongoDriver.java
@@ -48,6 +48,7 @@ import static io.vena.bosk.drivers.mongo.Formatter.REVISION_ONE;
 import static io.vena.bosk.drivers.mongo.Formatter.REVISION_ZERO;
 import static io.vena.bosk.drivers.mongo.Formatter.dottedFieldNameOf;
 import static io.vena.bosk.drivers.mongo.Formatter.enclosingReference;
+import static io.vena.bosk.drivers.mongo.v3.MainDriver.COLLECTION_NAME;
 import static java.lang.String.format;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.bson.BsonBoolean.FALSE;
@@ -63,8 +64,6 @@ final class SingleDocumentMongoDriver<R extends StateTreeNode> implements MongoD
 	private final Reference<R> rootRef;
 	private final String echoPrefix;
 	private final AtomicLong echoCounter = new AtomicLong(1_000_000_000_000L); // Start with a big number so the length doesn't change often
-
-	static final String COLLECTION_NAME = "boskCollection";
 
 	SingleDocumentMongoDriver(Bosk<R> bosk, MongoClientSettings clientSettings, MongoDriverSettings driverSettings, BsonPlugin bsonPlugin, BoskDriver<R> downstream) {
 		validateMongoClientSettings(clientSettings);
