+++ bosk-mongo/src/test/java/io/vena/bosk/drivers/mongo/AbstractMongoDriverTest.java
@@ -14,6 +14,7 @@ import io.vena.bosk.Reference;
 import io.vena.bosk.SideTable;
 import io.vena.bosk.annotations.ReferencePath;
 import io.vena.bosk.drivers.mongo.MongoDriverSettings.MongoDriverSettingsBuilder;
+import io.vena.bosk.drivers.mongo.v3.MainDriver;
 import io.vena.bosk.drivers.state.TestEntity;
 import io.vena.bosk.exceptions.InvalidTypeException;
 import java.lang.reflect.Method;
@@ -28,7 +29,6 @@ import org.junit.jupiter.api.TestInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static io.vena.bosk.drivers.mongo.SingleDocumentMongoDriver.COLLECTION_NAME;
 import static java.util.concurrent.TimeUnit.SECONDS;
 
 abstract class AbstractMongoDriverTest {
@@ -58,7 +58,7 @@ abstract class AbstractMongoDriverTest {
 		// Start with a clean slate
 		mongoService.client()
 			.getDatabase(driverSettings.database())
-			.getCollection(COLLECTION_NAME)
+			.getCollection(MainDriver.COLLECTION_NAME)
 			.drop();
 	}
