+++ bosk-mongo/src/test/java/io/vena/bosk/drivers/mongo/MongoDriverConformanceTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 
-import static io.vena.bosk.drivers.mongo.SingleDocumentMongoDriver.COLLECTION_NAME;
+import static io.vena.bosk.drivers.mongo.v3.MainDriver.COLLECTION_NAME;
 
 @UsesMongoService
 class MongoDriverConformanceTest extends DriverConformanceTest implements TestParameters {
